test: adding tests for shelter-csv-importer
