+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "ai-test.endpoint": "a",
+  "ai-test.apiKey": "a"
+}
