+++ package.json
@@ -43,6 +43,7 @@
     "zod": "^3.23.6"
   },
   "devDependencies": {
+    "@anatine/zod-mock": "^3.13.4",
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
