+++ src/shelter-csv-importer/shelter-csv-importer.module.ts
@@ -1,4 +1,9 @@
 import { Module } from '@nestjs/common';
+import { PrismaModule } from '../prisma/prisma.module';
+import { ShelterCsvImporterService } from './shelter-csv-importer.service';
 
-@Module({})
+@Module({
+  imports: [PrismaModule],
+  providers: [ShelterCsvImporterService],
+})
 export class ShelterCsvImporterModule {}
