+++ test/examples/gemini_prompt_response_example.json
@@ -1,0 +1,291 @@
+{
+  "medicamentos": [
+    "Kits de primeiros socorros",
+    "Gaze",
+    "Termômetro",
+    "Álcool",
+    "Repelente",
+    "Soro",
+    "Remédios básicos",
+    "Medicamentos prescritos",
+    "Pomadas",
+    "Azitotromicina",
+    "Antibiótico",
+    "Álcool em gel",
+    "Pomada para assadura",
+    "Desinfetante",
+    "Medicamentos veterinários",
+    "Seringa De Insulina",
+    "Seringa 3 E 5ml Com Agulha 70x30",
+    "Hipoglós",
+    "Antipulgas - Animais Pequeno Porte",
+    "Probiótico Para Animais"
+  ],
+  "cuidados com animais": [
+    "Ração para animais",
+    "Caixa de areia (gato)",
+    "Voluntários Para Animais (não Necessariamente Veterinário/a)",
+    "Coleiras",
+    "Patê Para Cachorro",
+    "Casinha para cachorro",
+    "Lar Temporário Para Animais",
+    "Caixas de transporte para pets",
+    "Leitos para animais",
+    "SACOS DE AREIA",
+    "Caixa De Transporte",
+    "Areia de gato",
+    "TELAS / CHIQUEIROS",
+    "Tapetes higiênicos",
+    "Ração gato"
+  ],
+  "especialistas e profissionais": [
+    "Veterinários",
+    "Voluntário - Noite",
+    "Turnos de manhã",
+    "Turnos de noite",
+    "Turno da madrugada",
+    "Voluntário - Madrugada",
+    "Voluntário - Manhã",
+    "Médicos",
+    "Psicólogos",
+    "Assistentes Sociais",
+    "Turnos de tarde",
+    "Voluntário - Tarde",
+    "Técnico De Enfermagem",
+    "Enfermeiros"
+  ],
+  "acomodações e descanso": [
+    "Lençóis",
+    "Cadeiras (tipo de praia)",
+    "Colchões",
+    "Lonas",
+    "Travesseiros/almofada",
+    "Rolo Lona Preta 40m x 6m (mais grossa)",
+    "Lona",
+    "Cordas",
+    "Cama Geriatria",
+    "Cobertores",
+    "Guarda-sol",
+    "Travesseiro",
+    "Luminárias com pedestal para área saúde 1m altura pelo menos",
+    "Ganchos para montar varal",
+    "Bermudas",
+    "Capas de chuva",
+    "Guarda-chuva",
+    "Roupas de Camas"
+  ],
+  "equipamentos de emergência": [
+    "Lanternas",
+    "Luz de Emergência",
+    "Lanterna",
+    "Ventilador",
+    "Ventiladores",
+    "Caminhão Pipa",
+    "Fita Durex larga",
+    "Cadeira De Rodas",
+    "Pilha Aa",
+    "Baldes"
+  ],
+  "voluntariado": [
+    "Voluntário - Noite",
+    "Turnos de manhã",
+    "Turnos de noite",
+    "Turno da madrugada",
+    "Voluntário - Madrugada",
+    "Voluntário - Manhã",
+    "Turnos de tarde",
+    "Voluntário - Tarde"
+  ],
+  "itens descartáveis": [
+    "Pratos",
+    "Descarpack Caixa",
+    "Mamadeira",
+    "Chupeta/bico",
+    "Fraldas",
+    "Sacos de lixo",
+    "Sacolinhas plasticas para kits",
+    "Jornal (para xixi e coco)",
+    "Copos",
+    "Papagaio - Urinar",
+    "Colher Descartável",
+    "Papel toalha",
+    "Guardanapo de papel",
+    "Embalagem descartável para as marmitas",
+    "Garfo e faca",
+    "Potes (marmitex)",
+    "Talher Descartável",
+    "Luva de Latex Descartável",
+    "Escova Para Limpar Mamadeira",
+    "Toucas",
+    "Luvas Descartáveis",
+    "Lenço umedecido",
+    "Assento para vaso sanitário",
+    "Jornal/Papelão",
+    "Touca Descartável",
+    "Sacolas Para Montar Kits",
+    "Sacolas ou sacos plásticos"
+  ],
+  "higiene pessoal": [
+    "Absorventes higiênicos",
+    "Lenços umedecidos",
+    "Sabonetes",
+    "Escova de dentes",
+    "Shampoo e Condicionador",
+    "Pás",
+    "Desodorante",
+    "Shampoo",
+    "Roupas íntimas",
+    "Roupa Íntima (adulto) Feminina E Masculina",
+    "Toalhas de banho",
+    "Papel higiênico",
+    "Roupa Íntima Infantil",
+    "Sabão em pó",
+    "Pasta de dente",
+    "Produtos de desinfecção (cloro, álcool)",
+    "Escova de cabelo",
+    "Toalhas",
+    "Pente/Escova de cabelo",
+    "Condicionador",
+    "Gilete",
+    "Tapete higiênico",
+    "Sabonete infantil",
+    "Higiene Pessoal",
+    "Produtos de higiene",
+    "Pano de chão",
+    "Pano de prato",
+    "Detergente",
+    "Água Sanitária",
+    "Chimia",
+    "Sabonete Líquido",
+    "Pano De Chão / Saco Alvejado",
+    "Esponjas De Louça",
+    "Lenço umedecido",
+    "Luvas para limpeza",
+    "Produtos de limpeza",
+    "Pano De Limpeza",
+    "Creme de pentear",
+    "Toalha de Banho",
+    "Itens de higiena pessoal"
+  ],
+  "alimentos e água": [
+    "Roupas para crianças",
+    "Comidas Não Perecíveis",
+    "Fórmula infantil",
+    "Carne Gado",
+    "Massa",
+    "Feijão",
+    "Oléo",
+    "Leite em pó",
+    "Gelo",
+    "Suco de caixinha",
+    "Alimentos Diets",
+    "Água",
+    "Chinelos masculinos",
+    "Água potável",
+    "Azeite",
+    "Arroz",
+    "Macarrão",
+    "Café",
+    "Alimentos para consumo rápido (Leite, bolacha, suco, etc)",
+    "Lanches",
+    "Pão",
+    "Pão De Sanduíche",
+    "Sucos",
+    "Banana",
+    "Cebola",
+    "Frutas",
+    "Alface",
+    "Tomate",
+    "Leite",
+    "Molho de tomate",
+    "Açúcar",
+    "Frios",
+    "Carne Frango",
+    "Farinha",
+    "Bolachinha",
+    "Farinha de trigo",
+    "Gatorade/Isotônico",
+    "Massinha De Modelar",
+    "Sal",
+    "Polenta",
+    "Achocolatado pronto"
+  ],
+  "material de limpeza": [
+    "Caixas de papelão",
+    "Lixeiras",
+    "Luvas De Limpeza",
+    "Vassouras",
+    "Sacos de lixo",
+    "Itens de limpeza",
+    "Produtos de desinfecção (cloro, álcool)",
+    "Desinfetante",
+    "Vassouras e rodos",
+    "Latas de lixo",
+    "Saco De Lixo De Vários Tamanhos"
+  ],
+  "vestuário": [
+    "Roupas para crianças",
+    "Plus Size/GG",
+    "Luvas",
+    "Roupas para frio",
+    "Roupas Masculinas G E Gg",
+    "Roupa íntima masculina e feminina",
+    "Roupas femininas G/GG",
+    "Roupa plus size",
+    "Roupas grandes",
+    "Roupas para adultos",
+    "Sapato Infantil",
+    "Roupas Femininas",
+    "Roupa Masculina",
+    "Roupas Pluz Size",
+    "Fraldas Geriatricas",
+    "Mascara",
+    "Roupa Masculina Gg",
+    "Roupa Infantil Menino",
+    "Roupas INTIMAS - CUECAS - G",
+    "Bermuda Masculina",
+    "Roupa Infantil 8-10 Anos",
+    "Roupa Masculina Adulta",
+    "Fralda P",
+    "Fraldas Adultas",
+    "Meias",
+    "Roupas leves",
+    "Bermudas",
+    "Calçado masculino",
+    "Roupas plus size feminina",
+    "Bonés",
+    "Roupa íntima feminina",
+    "Fralda XG e XXG",
+    "Roupas plus size masculino",
+    "Calçados",
+    "Roupa Íntima Infantil"
+  ],
+  "veículos de resgate e transporte": [
+    "Caminhão Pipa"
+  ],
+  "eletrodomésticos e eletrônicos": [
+    "Ventilador",
+    "Ventiladores",
+    "Luz de Emergência",
+    "Lanterna",
+    "Freezer para armazenar comida",
+    "Máquina de lavar roupa",
+    "Luminárias com pedestal para área saúde 1m altura pelo menos"
+  ],
+  "mobílias": [
+    "Cadeiras (tipo de praia)",
+    "Colchões",
+    "Cama Geriatria"
+  ],
+  "jogos e passatempo": [
+    "Jogo de cartas, baralho, dama e dominó, etc",
+    "Baralhos"
+  ],
+  "cosméticos": [
+    "Shampoo e Condicionador",
+    "Pás",
+    "Desodorante",
+    "Shampoo",
+    "Creme de pentear"
+  ]
+}
