+++ test/examples/sheet_example.csv
@@ -1,0 +1,7 @@
+nome_do_local,endereco,whatsapp,lat,lng,itens_disponiveis,itens_em_falta
+Centro Humanista,"Av. baltazar de oliveira garcia, 2132 - Porto Alegre",https://chat.whatsapp.com/Fs967Q6vvIn6F77pBUTnyF,-300.106.249,-511.227.617,Água potável,"Lanternas, Roupas para crianças, Comidas Não Perecíveis, Pratos, Lençóis, Cadeiras (tipo de praia), Absorventes higiênicos, Kits de primeiros socorros, Fórmula infantil, Colchões"
+Igreja Hebrom - Canoas,Av Esperança 1560,51 992726416,-2.989.020.757.796.020,-5.114.125.020.208.300,Água potável,"Caixas de papelão , Lixeiras, Carne Gado, Gaze, Termômetro, luvas, Álcool, Repelente, Massa, Descarpack Caixa"
+Igreja Reviver,"Av. Edgar Pires De Castro, 803 - Hípica",(51) 98493-2401,-301.566.507,-511.838.467,Água potável,"Pomadas, Feijão , Tênis, Chinelos, Oléo, Roupas para frio, Leite em pó, Vassouras, Luvas, Roupas íntimas"
+Associação Moradores Jardim Luciana - São Leopoldo,"Rua Tio Tiete, 510 - Jardim Luciana",,-2.970.870.574,-5.119.128.243,Água potável,"Medicamentos prescritos, Comidas Não Perecíveis, Turnos de manhã, Pratos, Lençóis, Absorventes higiênicos, Turnos de noite, Fraldas, Fórmula infantil, Colchões"
+Paróquia Santa Clara,"Estrada João De Oliveira Remião, 4444 - Lomba Do Pinheiro",,-301.081.531,-51.112.884,Água potável,"Comidas Não Perecíveis, Ventilador, Roupa Íntima (adulto) Feminina E Masculina, Luvas De Limpeza, Toalhas de banho, Ventiladores, Repelente, Vassouras, Massa, Sacos de lixo"
+Usina do Gasômetro,"na praça em frente ao Gasômetro, astrás do parque Harmonia",,,,Água potável,"Gelo, Azitotromicina, Papel higiênico, Absorventes higiênicos, Chinelos, Sacolinhas plasticas para kits , Gás, Ração para animais, Antibiótico , Suco de caixinha "
